Add in-memory session storage
Demo app was using FileStore to keep a file for every session created
with the app, which is, basically, a memory leak for the service. A new
in-memory store was added to it and now it's possible to flush all the
expired sessions after an amount of time. Each session is kept alive
during 30min and flushed from memory after every 24h.